+++ src/Functions/Functions.Autorest/Properties/AssemblyInfo.cs
@@ -33,3 +33,4 @@ using System.Runtime.InteropServices;
 
 
 
+